refactor: making theme changer addon working again
