+++ packages/core/.storybook/decorators/Frame.jsx
@@ -2,10 +2,17 @@ import React from 'react';
 import '../../components/GlobalStyle/GlobalStyle.css';
 import * as themes from '../../components/ThemeProvider/theme';
 
-const Frame = Story => (
-  <div style={{ padding: 10 }} className={themes.win95}>
-    <Story />
-  </div>
-);
+const Frame = (Story, { globals }) => {
+  const { selectedTheme } = globals;
+
+  return (
+    <div
+      style={{ padding: 10 }}
+      className={themes[selectedTheme] || themes.win95}
+    >
+      <Story />
+    </div>
+  );
+};
 
 export default Frame;
