+++ packages/core/components/ThemeProvider/theme.ts
@@ -2,4 +2,4 @@ import { win95 } from './themes/win95.css';
 import { bee } from './themes/bee.css';
 import { counterStrike } from './themes/counterStrike.css';
 
-export { win95, bee, counterStrike as counterstrike };
+export { win95, bee, counterStrike };
